backend/requirements.txt
# Python dependencies with pre-built wheels for Render compatibility
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
firebase-admin==6.4.0
python-dotenv==1.0.0
pydantic==2.8.2
python-multipart==0.0.6
nltk==3.8.1
resend==0.8.0
pywebpush==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
email-validator==2.1.0
# Scalability dependencies
slowapi==0.1.9
redis==5.0.1
